// frontend/src/generated/types/paymentCouponRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */
import type { PaymentCoupon } from './paymentCoupon';

export interface PaymentCouponRequest {
  newItems?: PaymentCoupon[];
  updateItems?: PaymentCoupon[];
  delItems?: PaymentCoupon[];
}